[WIP] Add decimal timer with separate thread (#13)
* Add decimal timer with separate thread

* Restart counter when the game is restarted

* Make the game randomise steps on new round

+++ app/src/main/java/com/tomaszrykala/dorefindmi/things/supplier/digidisplay/DigiDisplaySupplier.java
@@ -4,7 +4,7 @@ import com.tomaszrykala.dorefindmi.things.supplier.BaseSupplier;
 
 public interface DigiDisplaySupplier extends BaseSupplier {
 
-    void display(int n);
+    void display(double n);
 
     void display(String s);
 }
